Register "File>Import>MIDI..."; break FileMenus -> ImportMIDI...
... Though one might question whether wexreally need this item, distinct from
Import Audio.
